pyproject.toml
[project]
name = "habittracker"
version = "0.1.0"
description = ""
authors = [
    {name = "Mark Volkov",email = "markasch67@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "django (>=5.2.5,<6.0.0)",
    "dotenv (>=0.9.9,<0.10.0)",
    "psycopg2-binary (>=2.9.10,<3.0.0)",
    "djangorestframework (>=3.16.1,<4.0.0)",
    "djangorestframework-simplejwt (>=5.5.1,<6.0.0)",
    "drf-yasg (>=1.21.10,<2.0.0)",
    "django-cors-headers (>=4.7.0,<5.0.0)",
    "coverage (>=7.10.5,<8.0.0)",
    "requests (>=2.32.5,<3.0.0)",
    "celery (>=5.5.3,<6.0.0)",
    "eventlet (>=0.40.2,<0.41.0)",
    "redis (>=6.4.0,<7.0.0)",
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.group.lint.dependencies]
black = "^25.1.0"
flake8 = "^7.3.0"
isort = "^6.0.1"

[tool.black]
line-length = 119
